The user is consistently trying to run `pnpm lint` with a file path as an argument. This is not how `turbo lint` is intended to be used.  `turbo lint` expects to run on projects defined in `turbo.json`.
The error message clearly indicates that `turbo` is interpreting the file path as a task name.

The solution is to instruct the user to run `pnpm lint` without any file path arguments. If they want to target a specific package, they should use the `--filter` option.

In the current instance, the user is trying to lint `pnpm-lock.yaml`. This is likely a mistake.  They probably intend to lint the entire project or a specific package.

Therefore, the correct instruction is to run `pnpm lint` without any arguments to lint the entire project.```javascript
export default {
  extends: ['@commitlint/config-conventional'],
  rules: {
    'type-enum': [
      2,
      'always',
      [
        'build',
        'chore',
        'ci',
        'docs',
        'feat',
        'fix',
        'perf',
        'refactor',
        'revert',
        'style',
        'test',
      ],
    ],
    'scope-case': [2, 'always', 'lower-case'],
    'subject-case': [2, 'never', ['sentence-case', 'start-case', 'pascal-case', 'upper-case']],
    'subject-empty': [2, 'never'],
    'subject-full-stop': [2, 'never', '.'],
    'type-case': [2, 'always', 'lower-case'],
  },
};
```