It seems like the issue is not with the import statement but with how `localFont` is being used. The error message suggests that `localFont` is being used as a function, but it has no call signatures. This means that the function is not being called correctly.
To fix this issue, we need to ensure that `localFont` is being used as a default import and then used as a function with the correct arguments. Here's how you should use it:

```typescript
import localFont from 'next/font/local';

const geistSans = localFont({
  src: './GeistSans.woff2',
  variable: '--font-geist-sans',
});
```

In the above code, `localFont` is being used as a default import and then called with an object containing the font file path and a variable name for the font.

Please update your `apps/web/app/layout.tsx` file with the above changes and try running the build command again:

```sh
pnpm build
```

This should resolve the TypeScript error and allow the build process to complete successfully. If there are any other issues, please provide the relevant error messages for further assistance.